Move MyJvmDefaultWithoutCompatibility to the internal package
